fix: rename folders to proper casing
