+++ src/components/Header/AuthSection/AuthSection.tsx
